refactor: consolidate image replacement configuration schemas
