+++ src/config/schema.ts
@@ -3,6 +3,7 @@ import {
   type StringifyMarkdownOptions,
   StringifyMarkdownOptionsSchema,
 } from '@vivliostyle/vfm';
+import type * as mupdfType from 'mupdf';
 import { satisfies as semverSatisfies } from 'semver';
 import upath from 'upath';
 import * as v from 'valibot';
@@ -10,7 +11,6 @@ import * as v from 'valibot';
 import { CONTAINER_URL } from '../constants.js';
 import type { LoggerInterface } from '../logger.js';
 import { cliVersion } from '../util.js';
-import type { ReplaceFunction } from './replace-image.js';
 
 const $ = (strings: TemplateStringsArray, ...values: unknown[]) => {
   const lines = String.raw({ raw: strings }, ...values).split('\n');
@@ -355,6 +355,34 @@ const CmykSchema = v.pipe(
   `),
 );
 
+/** Values available while a replacement function is running. */
+export interface ReplaceFunctionContext {
+  /**
+   * The current PDF image as an owned reference scoped to this invocation.
+   * Its ownership is moved into the replacement function and Vivliostyle CLI
+   * destroys it when the function settles unless it is returned as the
+   * replacement. It must not be destroyed manually or retained after the
+   * function settles. Native objects returned by its methods are owned by the
+   * replacement function and must be destroyed before it settles unless that
+   * object is an image returned as the replacement.
+   */
+  image: mupdfType.Image;
+  /** The MuPDF module that owns the current image. */
+  mupdf: typeof import('mupdf');
+}
+
+/**
+ * Returns an owned replacement image, transferring its ownership to
+ * Vivliostyle CLI, or `null` to decline the current match and continue to the
+ * next replacement candidate. The returned image must be created with the
+ * supplied `mupdf` module and must not be used or destroyed after it is
+ * returned. The current `image` may be returned directly to use it as the
+ * replacement.
+ */
+export type ReplaceFunction = (
+  context: ReplaceFunctionContext,
+) => mupdfType.Image | null | Promise<mupdfType.Image | null>;
+
 const ReplaceFunctionSchema = v.pipe(
   // oxlint-disable-next-line typescript/no-unsafe-type-assertion
   v.function() as v.GenericSchema<ReplaceFunction>,
@@ -364,7 +392,7 @@ const ReplaceFunctionSchema = v.pipe(
   ),
 );
 
-const ImageConversionReplacementSchema = v.pipe(
+export const ImageConversionReplacementSchema = v.pipe(
   v.variant('kind', [
     v.object({
       kind: v.literal('builtin'),
@@ -388,6 +416,9 @@ const ImageConversionReplacementSchema = v.pipe(
   v.title('ImageConversionReplacement'),
   v.description('Image color conversion created by a replacement factory.'),
 );
+export type ImageConversionReplacement = Readonly<
+  v.InferInput<typeof ImageConversionReplacementSchema>
+>;
 
 const ReplaceImageEntrySchema = v.pipe(
   v.object({
@@ -410,6 +441,7 @@ const ReplaceImageEntrySchema = v.pipe(
   }),
   v.title('ReplaceImageEntry'),
 );
+export type ReplaceImageEntry = v.InferInput<typeof ReplaceImageEntrySchema>;
 
 const ReplaceImageSchema = v.pipe(
   v.array(
@@ -426,6 +458,31 @@ const ReplaceImageSchema = v.pipe(
     replaceable image.
   `),
 );
+export type ReplaceImageConfig = v.InferInput<typeof ReplaceImageSchema>;
+
+export interface ResolvedReplaceFunction {
+  replaceFunction: ReplaceFunction;
+  label: string;
+}
+
+export interface ResolvedImageConversionReplacement {
+  imageConversion: ImageConversionReplacement;
+  label: string;
+}
+
+export type ResolvedReplacement =
+  | ResolvedReplaceFunction
+  | ResolvedImageConversionReplacement;
+
+export interface ResolvedReplaceImageEntry {
+  source: string;
+  replacement: string | ResolvedReplacement;
+}
+
+export type ResolvedReplaceImageConfig = (
+  | ResolvedReplaceImageEntry
+  | ResolvedReplacement
+)[];
 
 const PdfPostprocessConfigSchema = v.pipe(
   v.partial(
